add README

+++ src/database/migrations/2024_12_23_120226_create_orders_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->foreignId('item_id')->constrained()->cascadeOnDelete();
             $table->foreignId('buyer_id')->constrained('users')->cascadeOnDelete();
             $table->foreignId('seller_id')->constrained('users')->cascadeOnDelete();
-            $table->integer('status')->default(0);
+            $table->tinyInteger('status')->default(0);
             $table->string('shipping_postal_code');
             $table->string('shipping_address');
             $table->string('shipping_building');
